Start forcibly loading the Ticket object we use in Scrips
We've done several things with this in the past.
Take an alias to the TicketObj (which means changing CurrentUser leaks).
Clone the object, which means that TransactionBatch Scrips run twice.

This code forcibly loads a clean TicketObj for the use of the Scrip
system.  It then removes the _TransactionBatch cache from tickets being
processed from ApplyTransactionBatch.

It does not re-process the source objects on Commit, so likely we need a
bit more poking to make sure no code is just invoking Commit with
non-superuser Objects.

At the end of the Commit, it removes the _TransactionBatch objects from
the copied ticket to ensure that we don't run TransactionBatch twice.

We really need a way to say "Run transaction batch, for these objects"
and not copy around an arrayref of Transaction Objects as a boolean
flag.